Fix PicoScope streaming acquisition cancellation race condition.
Fixed cancellation race condition on streaming acquisition by starting the workflow with its own cancellation token. No other commands to the instrument should be registered on the provided cancellation token. See the updated Demo.fsx for an example of how to achieve this. Logically speaking, this makes sense: the workflow which is opening and closing the connection to the hardware is not being cancelled. Although the code is quite complex, this fixes #33.