Complete block and buffer signals in PicoScopes
Block and buffered data modes are no longer second-class citizens of the
PicoScope signally API, with all modules exposing the same functions.

To achieve this, I generalised everything into curried functions, so while
there's still some code duplication between the modules, the vast majority of
the duplication that existed before is now gone, so the whole thing should be
more extensible.

If we need to expose the extensibility in the future, we can just remove the
private nature of the generic functions that are in each module.