Add multiple block acquisition
This is very hacky, and uses a totally different method to every other kind of
acquisition, but it works and it produces blocks, so it's good enough for us
now, pending a more complete rewrite of the acquisition code.

Nothing has been removed, only additions - all existing functionality will
continue working exactly as it did before.